Gemini/requirements.txt
mcp>=1.2.0
google-genai>=0.3.0
